# backend/swagger.yaml
openapi: 3.0.3
info:
  title: API del Restaurante
  version: "1.0.0"
servers:
  - url: "http://localhost:3000"
paths:
...
  /order:
    get:
      summary: Listar órdenes
      responses:
        '200':
          description: OK
    post:
      summary: Crear orden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                staffid:
                  type: integer
                date:
                  type: string
                  format: date
      responses:
        '201':
          description: Creado
  /order/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Obtener orden
      responses:
        '200':
          description: OK
    put:
      summary: Actualizar orden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                staffid:
                  type: integer
                date:
                  type: string
                  format: date
      responses:
        '200':
          description: Actualizado
    delete:
      summary: Eliminar orden
      responses:
        '200':
          description: Eliminado
  /orderdetail:
    get:
      summary: Listar detalles de orden
      responses:
        '200':
          description: OK
    post:
      summary: Crear detalle de orden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderid:
                  type: integer
                menuid:
                  type: integer
                itemid:
                  type: integer
                quantity:
                  type: integer
      responses:
        '201':
          description: Creado
  /orderdetail/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Obtener detalle de orden
      responses:
        '200':
          description: OK
    put:
      summary: Actualizar detalle de orden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderid:
                  type: integer
                menuid:
                  type: integer
                itemid:
                  type: integer
                quantity:
                  type: integer
      responses:
        '200':
          description: Actualizado
    delete:
      summary: Eliminar detalle de orden
      responses:
        '200':
          description: Eliminado
  /reports:
    get:
      summary: Listar reportes
      responses:
        '200':
          description: OK
    post:
      summary: Crear reporte
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                staffid:
                  type: integer
                date:
                  type: string
                  format: date
      responses:
        '201':
          description: Creado
  /reports/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Obtener reporte
      responses:
        '200':
          description: OK
    put:
      summary: Actualizar reporte
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                staffid:
                  type: integer
                date:
                  type: string
                  format: date
      responses:
        '200':
          description: Actualizado
    delete:
      summary: Eliminar reporte
      responses:
        '200':
          description: Eliminado
  /role:
    get:
      summary: Listar roles
      responses:
        '200':
          description: OK
    post:
      summary: Crear rol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rolename:
                  type: string
      responses:
        '201':
          description: Creado
  /role/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Obtener rol
      responses:
        '200':
          description: OK
    put:
      summary: Actualizar rol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rolename:
                  type: string
      responses:
        '200':
          description: Actualizado
    delete:
      summary: Eliminar rol
      responses:
        '200':
          description: Eliminado